+++ Sources/JSONSchema.swift
@@ -52,20 +52,25 @@ public struct Schema {
   }
 
   public func validate(_ data: Any) throws -> ValidationResult {
-    let validator = JSONSchema.validator(for: schema)
+    let validator = try JSONSchema.validator(for: schema)
     return try validator.validate(instance: data)
   }
 
   public func validate(_ data: Any) throws -> AnySequence<ValidationError> {
-    let validator = JSONSchema.validator(for: schema)
+    let validator = try JSONSchema.validator(for: schema)
     return try validator.validate(instance: data)
   }
 }
 
 
-func validator(for schema: [String: Any]) -> Validator {
-  guard let schemaURI = schema["$schema"] as? String else {
+func validator(for schema: [String: Any]) throws -> Validator {
+  guard schema.keys.contains("$schema") else {
+    // Default schema
     return Draft4Validator(schema: schema)
+  }
+
+  guard let schemaURI = schema["$schema"] as? String else {
+    throw ReferenceError.notFound
   }
 
   if let id = DRAFT_2020_12_META_SCHEMA["$id"] as? String, schemaURI == id {
@@ -84,11 +89,11 @@ func validator(for schema: [String: Any]) -> Validator {
     return Draft6Validator(schema: schema)
   }
 
-  if let id = DRAFT_04_META_SCHEMA["$id"] as? String, schemaURI == id {
+  if let id = DRAFT_04_META_SCHEMA["id"] as? String, schemaURI == id {
     return Draft4Validator(schema: schema)
   }
 
-  return Draft4Validator(schema: schema)
+  throw ReferenceError.notFound
 }